+++ src/components/AuthLogin/AuthModal.tsx
@@ -12,9 +12,11 @@ import {
   AUTH_WALLET_PROJECTS,
   saveTokenString,
   getUserIdFromJwtToken,
-  AUTH_WALLET_USER_PLAN,
 } from "../../lib/siwsrp/auth";
-import { DASHBOARD_URL, REQUEST_PARAMS } from "@site/src/lib/constants";
+import {
+  REQUEST_PARAMS,
+  AUTH_WALLET_USER_PLAN,
+} from "@site/src/lib/constants";
 import { MetamaskProviderContext } from "@site/src/theme/Root";
 
 Modal.setAppElement("#__docusaurus");
@@ -145,7 +147,7 @@ const AuthModal = ({
   setUksTier,
 }: AuthModalProps) => {
   const { siteConfig } = useDocusaurusContext();
-  const { DASHBOARD_PREVIEW_URL, VERCEL_ENV } = siteConfig?.customFields || {};
+  const { DASHBOARD_URL, VERCEL_ENV } = siteConfig?.customFields || {};
   const {
     sdk,
     setWalletLinked,
@@ -154,6 +156,7 @@ const AuthModal = ({
     setProjects,
     setMetaMaskAccount,
     setMetaMaskProvider,
+    fetchLineaEns,
   } = useContext(MetamaskProviderContext);
   const location = useLocation();
   const { pathname } = location;
@@ -167,9 +170,10 @@ const AuthModal = ({
 
       // Try to connect wallet first
       const accounts = await sdk.connect();
-      setMetaMaskAccount(accounts);
+
       if (accounts && accounts.length > 0) {
         setMetaMaskAccount(accounts[0]);
+        fetchLineaEns(accounts[0]);
         const provider = sdk.getProvider();
         setMetaMaskProvider(provider);
       }
@@ -182,7 +186,7 @@ const AuthModal = ({
 
       const loginResponse = await (
         await fetch(
-          `${DASHBOARD_URL(DASHBOARD_PREVIEW_URL, VERCEL_ENV)}/api/wallet/login`,
+          `${DASHBOARD_URL}/api/wallet/login`,
           {
             ...REQUEST_PARAMS("POST", {
               hydra_token: accessToken,
@@ -212,7 +216,7 @@ const AuthModal = ({
           }),
         ).toString("base64");
 
-        const walletLinkUrl = `${DASHBOARD_URL(DASHBOARD_PREVIEW_URL, VERCEL_ENV)}/login?mm_auth=${mm_auth}&redirect_to=${session.redirect_to}`;
+        const walletLinkUrl = `${DASHBOARD_URL}/login?mm_auth=${mm_auth}&redirect_to=${session.redirect_to}`;
 
         setWalletLinkUrl(walletLinkUrl);
 
@@ -248,7 +252,7 @@ const AuthModal = ({
 
       // You can use Infura Access Token to fetch any Infura API endpoint
       const projectsResponse = await fetch(
-        `${DASHBOARD_URL(DASHBOARD_PREVIEW_URL, VERCEL_ENV)}/api/v1/users/${userId}/projects`,
+        `${DASHBOARD_URL}/api/v1/users/${userId}/projects`,
         {
           ...REQUEST_PARAMS("GET", { Authorization: `Bearer ${token}` }),
         },
@@ -260,7 +264,7 @@ const AuthModal = ({
       setProjects(projects);
 
       const uksUserRawResp = await fetch(
-        `${DASHBOARD_URL(DASHBOARD_PREVIEW_URL, VERCEL_ENV)}/api/v1/users/${userId}`,
+        `${DASHBOARD_URL}/api/v1/users/${userId}`,
         {
           ...REQUEST_PARAMS("GET", { Authorization: `Bearer ${token}` }),
         },