+++ src/components/CodeTerminal/AlertMsg.jsx
@@ -1,6 +1,14 @@
 import React from "react";
-import { MSG_TYPES, DASHBOARD_URL } from "@site/src/lib/constants";
+import { MSG_TYPES } from "@site/src/lib/constants";
 import Link from "@docusaurus/Link";
+import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
+
+const getDashboardUrl = () => {
+  const { siteConfig } = useDocusaurusContext();
+  const { DASHBOARD_URL } = siteConfig?.customFields || {}
+  
+  return DASHBOARD_URL
+}
 
 export const INFO_MSG = {
   EMPTY_MSG: {
@@ -28,7 +36,7 @@ export const INFO_MSG = {
       <>
         Your request has been sent to the network endpoint you selected which
         returned a response with the requested information. Let’s see how it
-        performed in <Link to={`${DASHBOARD_URL}/dashboard/stats`}>stats</Link> or
+        performed in <Link to={`${getDashboardUrl}/dashboard/stats`}>stats</Link> or
         dive into{" "}
         <Link to="https://docs.infura.io/dashboard-stats">stats docs</Link>.
       </>
@@ -52,7 +60,7 @@ export const INFO_MSG = {
     description: (
       <>
         You don’t have any API Keys yet.{" "}
-        <Link to={`${DASHBOARD_URL}/dashboard`}>Create an API key</Link> to access
+        <Link to={`${getDashboardUrl}/dashboard`}>Create an API key</Link> to access
         your keys and send requests.
       </>
     ),
@@ -65,7 +73,7 @@ export const INFO_MSG = {
       <>
         The API key selected is not authorized for this specific product. To
         enable access, please add this product to your API key through the{" "}
-        <Link to={`${DASHBOARD_URL}/dashboard`}>Infura Dashboard</Link>
+        <Link to={`${getDashboardUrl}/dashboard`}>Infura Dashboard</Link>
       </>
     ),
   },