+++ src/lib/constants.ts
@@ -1,14 +1,3 @@
-export const PROD_APP_URL = "https://app.infura.io";
-export const STAGE_APP_URL = "https://infura-app-staging.vercel.app";
-export const DEV_APP_URL = "http://localhost:3000";
-
-export const DASHBOARD_URL = (DASHBOARD_PREVIEW_URL, VERCEL_ENV) =>
-  DASHBOARD_PREVIEW_URL
-    ? DASHBOARD_PREVIEW_URL
-    : VERCEL_ENV === "production"
-      ? PROD_APP_URL
-      : STAGE_APP_URL;
-
 export const REF_PATH = "/wallet/reference/new-reference";
 
 const TEST_TRANSACTIONS = {
@@ -461,9 +450,9 @@ export const REF_WALLET_PATH = "/wallet/reference/";
 export const REF_FAUCET_PATH = "/developer-tools/faucet/";
 
 export const REF_ALLOW_LOGIN_PATH = [
-    REF_SERVICES_PATH,
-    REF_WALLET_PATH,
-    REF_FAUCET_PATH,
+  REF_SERVICES_PATH,
+  REF_WALLET_PATH,
+  REF_FAUCET_PATH,
 ];
 
 export const REQUEST_PARAMS = (method = "POST", headers = {}) => ({
@@ -473,13 +462,22 @@ export const REQUEST_PARAMS = (method = "POST", headers = {}) => ({
     "Content-Type": "application/json",
     "Cache-Control": "no-cache",
     Pragma: "no-cache",
+    mode: 'cors',
     ...headers
   },
 });
 
-export const AUTH_WALLET_SESSION_NAME = "auth.wallet.session";
-export const AUTH_WALLET_TOKEN = "auth.wallet.token";
-export const AUTH_WALLET_PROJECTS = "auth.wallet.projects";
-export const LINEA_DEV_URL = "https://linea-mainnet.dev.infura.org";
 export const LINEA_PROD_URL = "https://linea-mainnet.infura.io";
+export const LINEA_DEV_URL = "https://linea-mainnet.dev.infura.org";
 export const LINEA_REQUEST_URL = LINEA_PROD_URL;
+export const AUTH_WALLET_ENS = "auth.wallet.ens";
+export const AUTH_WALLET_PROJECTS = "auth.wallet.projects";
+export const AUTH_WALLET_USER_PLAN = "auth.wallet.uksTier";
+
+export const getWalletEns = () => {
+  return sessionStorage.getItem(AUTH_WALLET_ENS);
+};
+
+export const getUksTier = (): string => {
+  return sessionStorage.getItem(AUTH_WALLET_USER_PLAN);
+};