+++ src/pages/developer-tools/faucet.tsx
@@ -15,10 +15,9 @@ import {
   AlertPastActivity,
 } from "@site/src/components/Faucet";
 import { useAlert } from "react-alert";
-import {MetamaskProviderContext} from "@site/src/theme/Root";
-
+import { MetamaskProviderContext } from "@site/src/theme/Root";
 import styles from "./faucet.module.scss";
-import { DASHBOARD_URL, REQUEST_PARAMS } from "@site/src/lib/constants";
+import { REQUEST_PARAMS } from "@site/src/lib/constants";
 import { AlertBalanceTooLow } from "@site/src/components/Faucet/Alerts";
 import {
   trackInputChangeForSegment,
@@ -35,7 +34,8 @@ export const LINEA_URL = "https://sepolia.lineascan.build/tx/";
 
 export default function Faucet() {
   const { siteConfig } = useDocusaurusContext();
-  const { userId, token, uksTier, metaMaskAccount } = useContext(MetamaskProviderContext);
+  const { userId, token, uksTier, metaMaskAccount, metaMaskAccountEns } =
+    useContext(MetamaskProviderContext);
   const alert = useAlert();
   const [transactions, setTransactions] = useState(DEFAULT_TRANSACTIONS_LIST);
   const [isLoading, setIsLoading] = useState(false);
@@ -44,7 +44,7 @@ export default function Faucet() {
   const [isLineaMaintenance, setIsLineaMaintenance] = useState(false);
   const [isSepoliaMaintenance, setIsSepoliaMaintenance] = useState(false);
   const [faucetBypassDomain, setFaucetBypassDomain] = useState(false);
-  const { DASHBOARD_PREVIEW_URL, VERCEL_ENV } = siteConfig?.customFields || {};
+  const { DASHBOARD_URL } = siteConfig?.customFields || {};
 
   const isLimitedUserPlan = uksTier === "core" && !faucetBypassDomain;
 
@@ -61,7 +61,7 @@ export default function Faucet() {
 
   const getTransactions = async () => {
     const sepolia = await fetch(
-      `${DASHBOARD_URL(DASHBOARD_PREVIEW_URL, VERCEL_ENV)}/api/faucets/sepolia/transactions`,
+      `${DASHBOARD_URL}/api/faucets/sepolia/transactions`,
       {
         ...REQUEST_PARAMS("GET", { Authorization: `Bearer ${token}` }),
       },
@@ -70,7 +70,7 @@ export default function Faucet() {
     setTransactionsForNetwork("sepolia", sepoliaData);
 
     const linea = await fetch(
-      `${DASHBOARD_URL(DASHBOARD_PREVIEW_URL, VERCEL_ENV)}/api/faucets/linea/transactions`,
+      `${DASHBOARD_URL}/api/faucets/linea/transactions`,
       {
         ...REQUEST_PARAMS("GET", { Authorization: `Bearer ${token}` }),
       },
@@ -84,7 +84,7 @@ export default function Faucet() {
     const address = walletAddress.trim();
     try {
       const faucetRawResponse = await fetch(
-        `${DASHBOARD_URL(DASHBOARD_PREVIEW_URL, VERCEL_ENV)}/api/faucets/${network}?address=${address}`,
+        `${DASHBOARD_URL}/api/faucets/${network}?address=${address}`,
         {
           ...REQUEST_PARAMS("POST", { Authorization: `Bearer ${token}` }),
           body: JSON.stringify({ dstAddress: address }),
@@ -175,10 +175,12 @@ export default function Faucet() {
   }, [userId, token]);
 
   useEffect(() => {
-    if (metaMaskAccount && !walletAddress) {
+    if (metaMaskAccountEns) {
+      setWalletAddress(metaMaskAccountEns);
+    } else if (metaMaskAccount) {
       setWalletAddress(metaMaskAccount);
     }
-  }, [metaMaskAccount]);
+  }, [metaMaskAccount, metaMaskAccountEns]);
 
   const tabItemContent = (network: "linea" | "sepolia") => {
     return (