+++ wallet-sidebar.js
@@ -270,6 +270,11 @@ const sidebar = {
                 },
                 {
                   type: "doc",
+                  label: "Create a simple Starknet dapp",
+                  id: "how-to/use-non-evm-networks/starknet/create-a-simple-starknet-dapp"
+                },
+                {
+                  type: "doc",
                   label: "Troubleshoot",
                   id: "how-to/use-non-evm-networks/starknet/troubleshoot"
                 },