+++ docs/whats-new.md
@@ -9,9 +9,19 @@ The latest major MetaMask documentation updates are listed by the month they wer
 For a comprehensive list of recent product changes, visit the "Release Notes" section at the bottom
 of the [MetaMask developer page](https://metamask.io/developer/).
 
+## November 2024
+
+- Added tutorial for
+  [creating a simple Starknet dapp](/wallet/how-to/use-non-evm-networks/starknet/create-a-simple-starknet-dapp).
+  ([#1656](https://github.com/MetaMask/metamask-docs/pull/1656))
+- Documented Snaps [`Card`](/snaps/features/custom-ui/#card) and [`Section`](/snaps/features/custom-ui/#section) UI components.
+  ([#1707](https://github.com/MetaMask/metamask-docs/pull/1707) and
+  [#1700](https://github.com/MetaMask/metamask-docs/pull/1700))
+
 ## October 2024
 
-- Documented support for [non-EVM networks](/wallet/how-to/use-non-evm-networks), starting with [Starknet](/wallet/how-to/use-non-evm-networks/starknet). ([#1509](https://github.com/MetaMask/metamask-docs/pull/1509))
+- Documented support for [non-EVM networks](/wallet/how-to/use-non-evm-networks), starting with [Starknet](/wallet/how-to/use-non-evm-networks/starknet).
+  ([#1509](https://github.com/MetaMask/metamask-docs/pull/1509))
 - Documented [Scroll](/services/reference/scroll) support.
   ([#1641](https://github.com/MetaMask/metamask-docs/pull/1641))
 - Updated [Wallet landing page](/wallet) and added [Connect to MetaMask](/wallet/connect) section